# contracts/openapi/public-v1.yaml
openapi: 3.1.0
info:
  title: Detection Scheduling Workbench API
  version: 1.0.0
  description: Human-approved scheduling API for an electrical-product testing center.
servers:
  - url: /api/v1
security:
  - OidcSessionCookie: []
paths:
...
  /schedule-previews/{preview_id}/explanation:
    post:
      operationId: explainSchedulePreview
      parameters: [ { $ref: '#/components/parameters/PreviewId' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleExplanationRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/ScheduleExplanationResponse' }, '503': { $ref: '#/components/responses/Problem' } }
  /schedule-previews/preflight:
    post:
      operationId: preflightScheduleDataQuality
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DataQualityPreflightRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/DataQualityPreflightResponse' }, '503': { $ref: '#/components/responses/Problem' } }
...
  /events/{event_id}/diagnose:
    post:
      operationId: diagnoseEvent
      parameters: [ { $ref: '#/components/parameters/EventId' }, { $ref: '#/components/parameters/IdempotencyKey' } ]
      responses: { '202': { $ref: '#/components/responses/DiagnosisResponse' }, '503': { $ref: '#/components/responses/Problem' } }
  /events/{event_id}/case-candidates:
    post:
      operationId: createExceptionCaseCandidate
      parameters: [ { $ref: '#/components/parameters/EventId' } ]
      responses: { '200': { $ref: '#/components/responses/ExceptionCaseCandidateResponse' }, '503': { $ref: '#/components/responses/Problem' } }
  /exception-case-candidates/{candidate_id}/submit:
    post:
      operationId: submitExceptionCaseCandidate
      parameters: [ { $ref: '#/components/parameters/CaseCandidateId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExceptionCaseCandidateSubmission' } } } }
      responses: { '201': { $ref: '#/components/responses/ExceptionCaseReviewResponse' }, '409': { $ref: '#/components/responses/Problem' } }
...
  /notification-drafts:
    post:
      operationId: createNotificationDraft
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationDraftRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/NotificationDraftResponse' }, '503': { $ref: '#/components/responses/Problem' } }
  /notification-drafts/{draft_id}/send:
    post:
      operationId: sendNotificationDraft
      parameters: [ { $ref: '#/components/parameters/NotificationDraftId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationDraftSendRequest' } } } }
      responses: { '202': { $ref: '#/components/responses/NotificationDeliveryResponse' }, '409': { $ref: '#/components/responses/Problem' } }
...
  /audit-logs/filter-suggestions:
    post:
      operationId: suggestAuditFilters
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuditFilterSuggestionRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/AuditFilterSuggestionResponse' }, '503': { $ref: '#/components/responses/Problem' } }
...
components:
  securitySchemes:
    OidcSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-public_session
      description: OIDC-authenticated BFF session cookie for same-origin browser requests. The BFF sets this cookie after the OIDC login flow with HttpOnly, Secure, and SameSite protections; the React client does not attach a Bearer token and the browser sends the cookie automatically. The BFF validates the OIDC-authenticated session and derives actor identity, center scope, and roles exclusively from verified claims.
  parameters:
...
    PreviewId: { name: preview_id, in: path, required: true, schema: { type: string } }
...
    EventId: { name: event_id, in: path, required: true, schema: { type: string } }
...
    CaseCandidateId: { name: candidate_id, in: path, required: true, schema: { type: string } }
    NotificationDraftId: { name: draft_id, in: path, required: true, schema: { type: string } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, minLength: 1 } }
  responses:
    Problem: { description: Problem details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    DiagnosisResponse: { description: Diagnosis, content: { application/json: { schema: { $ref: '#/components/schemas/Diagnosis' } } } }
    ScheduleExplanationResponse: { description: Deterministic preview explanation, content: { application/json: { schema: { $ref: '#/components/schemas/ScheduleExplanation' } } } }
    DataQualityPreflightResponse: { description: Deterministic data-quality preflight, content: { application/json: { schema: { $ref: '#/components/schemas/DataQualityPreflight' } } } }
    ExceptionCaseCandidateResponse: { description: Non-persistent exception case candidate, content: { application/json: { schema: { $ref: '#/components/schemas/ExceptionCaseCandidate' } } } }
    ExceptionCaseReviewResponse: { description: Submitted PostgreSQL review-queue record, content: { application/json: { schema: { $ref: '#/components/schemas/ExceptionCaseReview' } } } }
...
    NotificationDraftResponse: { description: Non-persistent enhanced notification body draft, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationDraft' } } } }
    NotificationDeliveryResponse: { description: Accepted deterministic notification delivery, content: { application/json: { schema: { $ref: '#/components/schemas/NotificationDelivery' } } } }
...
    AuditFilterSuggestionResponse: { description: Editable audit-filter suggestion, content: { application/json: { schema: { $ref: '#/components/schemas/AuditFilterSuggestion' } } } }
...
  schemas:
    Problem: { type: object, required: [type, title, status], properties: { type: { type: string }, title: { type: string }, status: { type: integer }, detail: { type: string }, instance: { type: string } } }
...
    Citation: { type: object, required: [standard_title, version, clause, page, content], properties: { standard_title: { type: string }, version: { type: string }, clause: { type: string }, page: { type: integer }, content: { type: string } } }
    Diagnosis: { type: object, required: [event_id, affected_order_ids, frozen_step_ids, sla_risks, affected_resources, evidence, resolved_case_ids, recommendations, evidence_gaps, confidence, tool_calls, degraded, memory_summary_status], properties: { event_id: { type: string }, affected_order_ids: { type: array, items: { type: string } }, frozen_step_ids: { type: array, items: { type: string } }, sla_risks: { type: array, items: { type: object } }, affected_resources: { type: array, items: { type: object } }, evidence: { type: array, items: { $ref: '#/components/schemas/Citation' } }, resolved_case_ids: { type: array, items: { type: string } }, recommendations: { type: array, items: { type: string } }, evidence_gaps: { type: array, items: { type: string } }, confidence: { type: string, enum: [high, medium, low, insufficient] }, tool_calls: { type: array, maxItems: 5, items: { type: string, enum: [get_event_snapshot, get_order_snapshot, get_resource_snapshot, get_schedule_snapshot, search_standards, search_resolved_cases] } }, degraded: { type: boolean }, memory_summary_status: { type: string, enum: [available, degraded, not_used] } } }
    ScheduleExplanationRequest: { type: object, required: [subject_type, subject_id], properties: { subject_type: { type: string, enum: [preview, order, step] }, subject_id: { type: string } } }
    ScheduleExplanation: { type: object, required: [preview_id, subject_type, subject_id, summary, constraint_reasons, tradeoffs, frozen_step_ids, blockers, evidence_available, degraded], properties: { preview_id: { type: string }, subject_type: { type: string, enum: [preview, order, step] }, subject_id: { type: string }, summary: { type: string }, constraint_reasons: { type: array, items: { type: string } }, tradeoffs: { type: array, items: { type: string } }, frozen_step_ids: { type: array, items: { type: string } }, blockers: { type: array, items: { type: object } }, fallback_reason: { type: [string, 'null'] }, evidence_available: { type: boolean }, degraded: { type: boolean } } }
    DataQualityPreflightRequest: { type: object, required: [scope], properties: { scope: { type: string, enum: [order, resource, schedule_preview] }, order_id: { type: string }, resource_ids: { type: array, items: { type: string } }, preview_id: { type: string } } }
    DataQualityFinding: { type: object, required: [code, severity, message, blocking], properties: { code: { type: string }, severity: { type: string, enum: [info, warning, error] }, message: { type: string }, field: { type: string }, blocking: { type: boolean }, suggestion: { type: [string, 'null'] } } }
    DataQualityPreflight: { type: object, required: [scope, status, findings, deterministic, explanation_available, degraded], properties: { scope: { type: string, enum: [order, resource, schedule_preview] }, status: { type: string, enum: [passed, blocked] }, findings: { type: array, items: { $ref: '#/components/schemas/DataQualityFinding' } }, deterministic: { type: boolean, const: true }, explanation_available: { type: boolean }, degraded: { type: boolean } } }
    ExceptionCaseCandidate: { type: object, required: [candidate_id, event_id, source_candidate_hash, summary, trigger, impact, disposition, outcome, tags, evidence, retention_until, status], properties: { candidate_id: { type: string, description: Opaque non-persistent signed candidate reference }, event_id: { type: string }, source_candidate_hash: { type: string }, summary: { type: string }, trigger: { type: string }, impact: { type: string }, disposition: { type: string }, outcome: { type: string }, tags: { type: array, items: { type: string } }, evidence: { type: array, items: { $ref: '#/components/schemas/Citation' } }, retention_until: { type: string, format: date-time }, status: { type: string, const: candidate } } }
    ExceptionCaseCandidateSubmission: { type: object, required: [source_candidate_hash, summary, trigger, impact, disposition, outcome, tags, retention_until], properties: { source_candidate_hash: { type: string }, summary: { type: string, minLength: 1 }, trigger: { type: string, minLength: 1 }, impact: { type: string, minLength: 1 }, disposition: { type: string, minLength: 1 }, outcome: { type: string, minLength: 1 }, tags: { type: array, items: { type: string } }, retention_until: { type: string, format: date-time } } }
    ExceptionCaseReview: { type: object, required: [id, event_id, status, version, submitted_at], properties: { id: { type: string }, event_id: { type: string }, status: { type: string, const: pending_review }, version: { type: integer }, submitted_at: { type: string, format: date-time } } }
...
    NotificationDraftRequest: { type: object, required: [notification_id], properties: { notification_id: { type: string }, instruction: { type: string, maxLength: 1000 } } }
    NotificationDraft: { type: object, required: [draft_id, notification_id, source_hash, title, body, degraded], properties: { draft_id: { type: string, description: Opaque non-persistent signed draft reference }, notification_id: { type: string }, source_hash: { type: string }, title: { type: string }, body: { type: string }, degraded: { type: boolean } } }
    NotificationDraftSendRequest: { type: object, required: [source_hash, body], properties: { source_hash: { type: string }, body: { type: string, minLength: 1, maxLength: 10000 } } }
    NotificationDelivery: { type: object, required: [id, draft_id, status, accepted_at], properties: { id: { type: string }, draft_id: { type: string }, status: { type: string, const: accepted }, accepted_at: { type: string, format: date-time } } }
...
    AuditFilterSuggestionRequest: { type: object, required: [query], properties: { query: { type: string, minLength: 1, maxLength: 1000 } } }
    AuditFilterSuggestion: { type: object, required: [original_query, filters, explanation, uncertainty, degraded], properties: { original_query: { type: string }, filters: { type: array, items: { type: object, required: [field, operator, value], properties: { field: { type: string, enum: [actor_id, action, entity_id, created_at] }, operator: { type: string, enum: [equals, contains, gte, lte] }, value: { type: string } } } }, explanation: { type: string }, uncertainty: { type: boolean }, degraded: { type: boolean } } }
...